Cleaner is now DashboardCleaner
