+++ Famicom/Components/Pages/DashboardCleaner.razor
@@ -1,5 +1,5 @@
 ﻿@page "/cleaner"
-@inherits CleanerBase
+@inherits DashboardCleaner
 
 <PageTitle>Cleaner Panel</PageTitle>
